Extract EditorXR-specific ListView modifications
