+++ Scripts/ListView/EditorXRListViewItem.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bbe2d6246867c254b89a5a7999f3e13d
+timeCreated: 1467071852
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
