+++ Scripts/ListView/EditorXRListViewItemNestedData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cf1208bcc184f73418bcebe1782c67b6
+timeCreated: 1472783144
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
