--- Scripts/ListView/ListViewItem.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: ed2552c8cbe6f364498c1061a98fa856
-timeCreated: 1467071852
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
